redis/worker/requirements.txt
pip
setuptools
numpy
pandas
mysql-connector-python
Celery
diskcache
redis